// cypress/enums/Data.ts
export enum baseURL {
    Url = 'https://www.ss.lv/'
}

//Header
export enum TopButtons {
    AddAD = 'Подать ОБъявление',
    MyADs = "Мои Объявления",
    Search = "Поиск",
    Bookmarks = 'Закладки'
}

//Heading on main page
export enum Heading {
    Jobs = 'Работа и бизнес',
    Transport = 'Транспорт',
    RealEstate = 'Недвижимость',
    Construction = 'Строительство',
    Electronics = 'Эдектротехника',
    Clothes = 'Одежда, обувь',
    ForHome = 'Для дома',
    Manufacturing = 'Производство',
    ForKids = 'Для детей',
    Animals = 'Животные',
    Agriculture = 'Сельское хозяйство',
    Rest = 'Отдых, увлечения'
}

//Transport titles
export enum Transport {
    Cars = 'Легковые авто'
}

//Cars pages
export enum Brand {
    Peugeot = 'Peugeot'
}

//Search page

export enum EngineType {
     Diesel = 'Дизель',
     Petrol = 'Бензин',
     Gas = 'Бензин/газ',
     Hybrid = 'Hybrid',
     EV = 'Электрический'
}

export enum  TransmissionType {
    Automatic = 'Автомат',
    Manual = 'Ручная'
}

export enum BodyType {
    SUV = 'Джип',
    Cabrio = 'Кабриолет',
    Coupe = 'Купе',
    Van = 'Микроавтобус',
    Minivan = 'Минивэн',
    Pickup = 'Пикап',
    Sedan = 'Седан',
    Wagon = 'Универсал',
    Hatchback = 'Хэтчбек'
}

export enum CarDealType {
    Selling = "Продают",
    Buying = "Покупают",
    Trading = "Меняют",
    Other = "Разное"
}

export enum CarSortBy {
    Brand = "Марка",
    Model = "Модель",
    Year = "Год",
    Volume = "Объём",
    Engine = "Двигатель",
    Milenage = "Пробег",
    Color = "Цвет",
    Price = "Цена"
}